+++ kitty/cocoa_window.m
@@ -258,6 +258,7 @@ PENDING(clear_screen, CLEAR_SCREEN)
 PENDING(clear_last_command, CLEAR_LAST_COMMAND)
 PENDING(reload_config, RELOAD_CONFIG)
 PENDING(toggle_macos_secure_keyboard_entry, TOGGLE_MACOS_SECURE_KEYBOARD_ENTRY)
+PENDING(macos_cycle_through_os_windows, MACOS_CYCLE_THROUGH_OS_WINDOWS)
 PENDING(toggle_fullscreen, TOGGLE_FULLSCREEN)
 PENDING(open_kitty_website, OPEN_KITTY_WEBSITE)
 PENDING(hide_macos_app, HIDE)
@@ -319,6 +320,7 @@ typedef struct {
     GlobalShortcut clear_terminal_and_scrollback, clear_screen, clear_scrollback, clear_last_command;
     GlobalShortcut toggle_macos_secure_keyboard_entry, toggle_fullscreen, open_kitty_website;
     GlobalShortcut hide_macos_app, hide_macos_other_apps, minimize_macos_window, quit;
+    GlobalShortcut macos_cycle_through_os_windows;
 } GlobalShortcuts;
 static GlobalShortcuts global_shortcuts;
 
@@ -336,7 +338,7 @@ cocoa_set_global_shortcut(PyObject *self UNUSED, PyObject *args) {
     else Q(clear_terminal_and_scrollback); else Q(clear_scrollback); else Q(clear_screen); else Q(clear_last_command);
     else Q(reload_config); else Q(toggle_macos_secure_keyboard_entry); else Q(toggle_fullscreen);
     else Q(open_kitty_website); else Q(hide_macos_app); else Q(hide_macos_other_apps);
-    else Q(minimize_macos_window); else Q(quit);
+    else Q(minimize_macos_window); else Q(quit); else Q(macos_cycle_through_os_windows);
 #undef Q
     if (gs == NULL) { PyErr_SetString(PyExc_KeyError, "Unknown shortcut name"); return NULL; }
     int cocoa_mods;
@@ -805,6 +807,8 @@ cocoa_create_global_menu(void) {
                           action:@selector(performZoom:)
                    keyEquivalent:@""];
     [windowMenu addItem:[NSMenuItem separatorItem]];
+    MENU_ITEM(windowMenu, @"Cycle Through OS Windows", macos_cycle_through_os_windows);
+    [windowMenu addItem:[NSMenuItem separatorItem]];
     [windowMenu addItemWithTitle:@"Bring All to Front"
                           action:@selector(arrangeInFront:)
                    keyEquivalent:@""];
@@ -919,6 +923,20 @@ cocoa_toggle_secure_keyboard_entry(void) {
     [k toggle];
     [[NSUserDefaults standardUserDefaults] setBool:k.isDesired forKey:@"SecureKeyboardEntry"];
 }
+
+void
+cocoa_cycle_through_os_windows(void) {
+    NSArray *windows = [NSApp orderedWindows];
+    if (windows.count < 2) return;
+
+    NSWindow *keyWindow = [NSApp keyWindow];
+    NSUInteger index = [windows indexOfObject:keyWindow];
+    NSUInteger nextIndex = (index + 1) % windows.count;
+
+    NSWindow *nextWindow = windows[nextIndex];
+    [nextWindow makeKeyAndOrderFront:nil];
+}
+
 
 void
 cocoa_hide(void) {